+++ CHANGELOG.md
@@ -7,6 +7,31 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+## [0.28.0] - 2026-04-05
+
+### Added
+
+- Battery status for shades
+
+## [0.27.0] - 2026-03-01
+
+### Added
+
+- Support for Pico Paddle.
+
+## [0.26.0] - 2025-11-08
+
+### Added
+
+- Support for Caseta's Smart-Away feature.
+- Support for Triathlon Tilt-only Wood Blinds.
+
+## [0.25.0] - 2025-08-11
+
+### Added
+
+- `BUTTON_STATUS_MULTITAP` constant
+
 ## [0.24.0] - 2025-03-09
 
 ### Added
@@ -344,7 +369,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Ability to interact with Caséta bridge using LIP over Telnet and LEAP over SSH.
 
-[unreleased]: https://github.com/gurumitts/pylutron-caseta/compare/v0.24.0...HEAD
+[unreleased]: https://github.com/gurumitts/pylutron-caseta/compare/v0.28.0...HEAD
+[0.28.0]: https://github.com/gurumitts/pylutron-caseta/compare/v0.27.0...v0.28.0
+[0.27.0]: https://github.com/gurumitts/pylutron-caseta/compare/v0.26.0...v0.27.0
+[0.26.0]: https://github.com/gurumitts/pylutron-caseta/compare/v0.25.0...v0.26.0
+[0.25.0]: https://github.com/gurumitts/pylutron-caseta/compare/v0.24.0...v0.25.0
 [0.24.0]: https://github.com/gurumitts/pylutron-caseta/compare/v0.23.0...v0.24.0
 [0.23.0]: https://github.com/gurumitts/pylutron-caseta/compare/v0.22.0...v0.23.0
 [0.22.0]: https://github.com/gurumitts/pylutron-caseta/compare/v0.21.1...v0.22.0